+++ templates/admin/tables/emails.php
@@ -1,4 +1,7 @@
 <?php
+
+use AnyComment\Helpers\AnyCommentRequest;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly
 }
@@ -13,7 +16,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 		<?php
 
 		$ids    = isset( $_POST['emails'] ) && ! empty( $_POST['emails'] ) ? $_POST['emails'] : null;
-		$action = isset( $_POST['action'] ) && ! empty( $_POST['action'] ) ? $_POST['action'] : null;
+		if ( AnyCommentRequest::post( 'action' ) !== '-1' ) {
+			$action = sanitize_text_field( $_POST['action'] );
+		} elseif ( AnyCommentRequest::post( 'action2' ) !== '-1' ) {
+			$action = sanitize_text_field( $_POST['action2'] );
+		}
 		if ( $action !== null && $ids !== null && $action === 'delete' ) {
 			if ( \AnyComment\Models\AnyCommentEmailQueue::deleted_all( 'ID', $ids ) ) {
 				$messages = '<div id="message" class="updated notice is-dismissible"><p>' . __( 'Selected emails were deleted.', 'anycomment' ) . '</p></div>';