+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 **Fixes:**
 
+- Fixed issue when table's bottom bulk actions did not work properly
+- Fixed issue when buld action only deleted one selected item out of multiple selected
 - Fixed synchronization issue with service when local database did not have comments
 - Fixed issue with emails, when they did not send or add to the list to be sent
 - Fixed notice warning in dashboard.php